requirements.txt
python-dotenv
requests
websockets
pydantic
pandas
matplotlib
mplfinance